+++ src/lib/components/v2/tasks/CustomEnvTask.svelte
@@ -139,9 +139,9 @@
 
 <div class="alert alert-warning">
 	<i class="bi bi-exclamation-triangle"></i>
-	Collecting tasks with a custom Python environment will use that environment for running the tasks.
-	Be careful about changing this environment, as that may break existing workflows. It is recommended
-	to use custom Python environments only during task development or when something needed for your environment
+	Collecting tasks with a custom Python environment will use that environment for running the tasks. Be
+	careful about changing this environment, as that may break existing workflows. It is recommended to
+	use custom Python environments only during task development or when something needed for your environment
 	building isn't supported in Fractal server yet. Collect the tasks with regular Fractal task collection
 	for production setups.
 </div>