+++ src/routes/auth/login/oauth2/+page.svelte
@@ -2,6 +2,8 @@
 	import { goto } from '$app/navigation';
 	import { page } from '$app/state';
 	import { onMount } from 'svelte';
+	import { resolve } from '$app/paths';
+
 	const message = $derived(page.data.message);
 
 	onMount(async () => {
@@ -9,7 +11,7 @@
 			if (sessionStorage) {
 				sessionStorage.setItem('temporary-message', message);
 			}
-			await goto('/');
+			await goto(resolve('/'));
 		}
 	});
 </script>