+++ src/routes/v2/admin/task-groups/[taskGroupId]/reset/+page.svelte
@@ -4,6 +4,7 @@
 	import { displayStandardErrorAlert, getAlertErrorFromResponse } from '$lib/common/errors';
 	import { PropertyDescription } from 'fractal-components';
 	import Modal from '$lib/components/common/Modal.svelte';
+	import { resolve } from '$app/paths';
 
 	let errorAlert = undefined;
 
@@ -90,7 +91,7 @@
 			const result = /** @type {import('fractal-components/types/api').TaskGroupActivityV2} */ (
 				await response.json()
 			);
-			await goto(`/v2/admin/task-groups/activities?activity_id=${result.id}`);
+			await goto(resolve(`/v2/admin/task-groups/activities?activity_id=${result.id}`));
 		} else {
 			errorAlert = displayStandardErrorAlert(
 				await getAlertErrorFromResponse(response),